+++ generators/app/templates/e2e/src/page-objects/__auth._login.po.ts
@@ -7,8 +7,9 @@ import { browser, element, by } from 'protractor';
 
 export class LoginPage {
 <%   if (props.ui === 'ionic') { -%>
-  usernameField = element(by.css('ion-input[formControlName="username"]'));
-  passwordField = element(by.css('ion-input[formControlName="password"]'));
+  // Custom locator to find an element within shadow DOM (by.deepCss() doesn't work here)
+  usernameField = element(by.shadowCss('ion-input[formControlName="username"]', 'input'));
+  passwordField = element(by.shadowCss('ion-input[formControlName="password"]', 'input'));
   loginButton = element(by.css('ion-button[type="submit"]'));
 <%   } else { -%>
   usernameField = element(by.css('input[formControlName="username"]'));
@@ -19,9 +20,9 @@ export class LoginPage {
   constructor() {
   }
 
-  login() {
-    this.usernameField.sendKeys('test');
-    this.passwordField.sendKeys('123');
-    this.loginButton.click();
+  async login() {
+    await this.usernameField.sendKeys('test');
+    await this.passwordField.sendKeys('123');
+    await this.loginButton.click();
   }
 }